Remove dependency on saved copies of names
Collection Tree since Omeka 2.0 stores its own copies of the names of
each collection, so it wouldn't have to change how it works much. This
is not reliable though: there are hooks to keep the collection tree name
copies up to date, but if the plugin isn't active it will miss
added/updated collections (and there are probably other similar cases).

Instead, switch displays to render the collection's title directly, and
for some other cases like dropdowns, rely on the dropdown options
already generated by the core Collections model.

(#35)